Clean code on menu and added users functions
